+++ opensrp-chw/src/main/java/org/smartregister/chw/dao/DefaultChwPNCDaoFlv.java
@@ -1,0 +1,8 @@
+package org.smartregister.chw.dao;
+
+import org.smartregister.chw.core.dao.AbstractDao;
+
+public abstract class DefaultChwPNCDaoFlv extends AbstractDao implements ChwPNCDao.Flavor {
+
+
+}
