+++ opensrp-chw/src/main/java/org/smartregister/chw/interactor/DefaultPncUpcomingServiceInteractorFlv.java
@@ -12,7 +12,7 @@ import org.smartregister.chw.core.dao.AlertDao;
 import org.smartregister.chw.core.rule.PNCHealthFacilityVisitRule;
 import org.smartregister.chw.core.utils.CoreConstants;
 import org.smartregister.chw.core.utils.VaccineScheduleUtil;
-import org.smartregister.chw.dao.PNCDao;
+import org.smartregister.chw.dao.ChwPNCDao;
 import org.smartregister.chw.domain.PNCHealthFacilityVisitSummary;
 import org.smartregister.chw.util.PNCVisitUtil;
 import org.smartregister.domain.Alert;
@@ -40,7 +40,7 @@ public class DefaultPncUpcomingServiceInteractorFlv implements PncUpcomingServic
 
     private void evaluateHealthFacility(List<BaseUpcomingService> serviceList) {
         // create data to display
-        PNCHealthFacilityVisitSummary summary = PNCDao.getLastHealthFacilityVisitSummary(memberObject.getBaseEntityId());
+        PNCHealthFacilityVisitSummary summary = ChwPNCDao.getLastHealthFacilityVisitSummary(memberObject.getBaseEntityId());
         if (summary != null) {
             PNCHealthFacilityVisitRule visitRule = PNCVisitUtil.getNextPNCHealthFacilityVisit(summary.getDeliveryDate(), summary.getLastVisitDate());
