+++ opensrp-chw/src/wcaro/java/org/smartregister/chw/dao/ChwPNCDaoFlv.java
@@ -1,0 +1,4 @@
+package org.smartregister.chw.dao;
+
+public class ChwPNCDaoFlv extends DefaultChwPNCDaoFlv {
+}
